Recording, replay, free mode
All kinect modes currently work.  Some of it is a bit thrown together
(just really wanted a working implementation), but I'm going to clean it
up a bit in the next few days.  Should get us through the wizard of oz
shit with a nice interface for us all to test in.