Improvements to DataTemplates, AppState object
DataTemplates now include bones.  Tried to get this working with an
ObservableCollection on the KinectModel in combination with a Bone class,
but it was becoming an enormous pain in the ass and not worth my time.  So
for now, just list all the bones in App's DataTemplate for a Skeleton.

AppState object introduced.  Can tack stuff on here as necessary for
keeping track of what's going on in the application.  For now, just the
kinect's state (yes this should probably be somehow encoded on the
KinectModel... perhaps, but for now, this works) and the kinect model.